refactor: #390 `Member` business logic을 CQRS pattern 적용하여 분리
